Finish

+++ ParkShark/Models/ViewModels/DetailParkingForm.cs
@@ -4,7 +4,7 @@
     {
         public int Parking { get; set; }
         public int HourlyRate { get; set; }
-        public int ParkingFee { get; set; }
+        public decimal ParkingFee { get; set; }
         public string Status { get; set; }
     }
 }
