+++ ParkShark/Models/ViewModels/DetailParkingView.cs
@@ -7,7 +7,7 @@
         public string LicensePlate { get; set; }
         public DateTime TimeEntry { get; set; }
         public int HourlyRate { get; set; }
-        public int ParkingFee { get; set; }
+        public decimal ParkingFee { get; set; }
         public string Status { get; set; }
     }
 }
